requirements.txt
browser-use>=0.1.40
playwright>=1.44.0
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0
apscheduler>=3.10.4
psycopg2-binary>=2.9.9
streamlit>=1.35.0
streamlit-autorefresh>=0.0.1
feedparser>=6.0.11
beautifulsoup4>=4.12.0
requests>=2.32.0
python-dotenv>=1.0.0
pandas>=2.0.0
scrapling>=0.3.0
